refactor: standardize file names and update imports
