+++ src/app/components/gpa-contact-details/gpa-contact-details-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { GPAContactDetailsComponent } from './g&pa-contact-details.component';
+import { GPAContactDetailsComponent } from './gpa-contact-details.component';
 
 const routes: Routes = [
   { path: '', component: GPAContactDetailsComponent }
