+++ src/app/components/gpa-contact-details/gpa-contact-details.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { GPAContactDetailsRoutingModule } from './g&pa-contact-details-routing.module';
-import { GPAContactDetailsComponent } from './g&pa-contact-details.component';
+import { GPAContactDetailsRoutingModule } from './gpa-contact-details-routing.module';
+import { GPAContactDetailsComponent } from './gpa-contact-details.component';
 import { SharedModule } from 'src/app/shared/shared.module';
 
 @NgModule({
