+++ src/app/components/inbox/inbox.component.ts
@@ -3,8 +3,8 @@ import { MatTableDataSource } from '@angular/material/table';
 import { MatSort } from '@angular/material/sort';
 import { MatPaginator } from '@angular/material/paginator';
 import { SelectionModel } from '@angular/cdk/collections';
-import { WorkItem } from '../../models/workItem.model';
-import { InboxItem } from 'src/app/models/inboxItem.model';
+import { WorkItem } from '../../models/work-item.model';
+import { InboxItem } from 'src/app/models/inbox-item.model';
 import { Router } from '@angular/router';
 import { CoreService } from 'src/app/services/core.service';
 import { FormControl } from '@angular/forms';
