+++ src/app/models/executive-users.model.ts
@@ -1,5 +1,5 @@
 
-import { SabMember } from "./sab-member";
+import { SabMember } from './sab-member.model';
 export class ExecutiveUsers {	
 
 	dceos:SabMember[];
