+++ src/app/models/manager.model.ts
