+++ src/app/models/observation-card.model.ts
