+++ src/app/models/work-item.model.ts
